Reuse code from CMode in ScribbleMode.
